Update dnf task file
With these updates the /etc/dnf/automatic.conf file will be created if
it does not exist, as seems to be the case.  These changes also take
advantage of the INI structure of /etc/dnf/automatic.conf to use
community.general.ini_file vice ansible.builtin.lineinfile.  This
makes the code cleaner, less fragile, and easier to understand.