Fix broken test code
I had previously attempted to preemptively prepare the Molecule test
code for the addition of Fedora 41 in #38, but I failed to realize
that the Fedora version is stored in host.system_info.release not
host.system_info.codename.  In fact, host.system_info.codename is
actually an empty string for Fedora.